% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_codynamics.R
\name{mod_codynamics_server}
\alias{mod_codynamics_server}
\title{Co-Dynamics Module Server}
\usage{
mod_codynamics_server(id, app_data)
}
\arguments{
\item{id}{Module namespace id}

\item{app_data}{Reactive data list from mod_data_server}
}
\description{
Co-Dynamics Module Server
}
