% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hedgeratios.R
\name{mod_hedgeratios_server}
\alias{mod_hedgeratios_server}
\title{Hedge Ratios Module Server}
\usage{
mod_hedgeratios_server(id, app_data)
}
\arguments{
\item{id}{Module namespace id}

\item{app_data}{Reactive data list from mod_data_server}
}
\description{
Hedge Ratios Module Server
}
